code/src/requirements.txt
google-cloud-storage
apache-beam[gcp]
google-api-python-client
oauth2client
google-cloud-monitoring
google-auth